Converted ImageQualityEstimator Class to FrameFilter
